requirements.txt
openai~=1.59.7
llama-index~=0.12.12
fastapi~=0.115.6
uvicorn~=0.34.0
requests~=2.32.3
python-dotenv~=1.0.1
